simplify dev database reset
